+++ packages/client/ui/react-ui/src/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useCrossmint";
 export * from "./useCrossmintCheckout";
 export * from "./useWallet";
 export * from "./useAuth";
+export * from "./useWalletCache";